Implement a distributed hashmap
The main bottleneck for the first version was the insertion in the
hashmap. To fix it, the implementation of the hashmap saw during the
lesson has been adapted to our usecase. The main difference is that a
node has now a value (the depth of the board). A successful insert is
made when a board is in the set and the value of its node is less than
the one already in the set (if any).